requirements.txt
fastapi==0.110.0
gunicorn==21.2.0
uvicorn==0.27.1
google-genai==0.3.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.28
alembic==1.13.1
redis==5.0.3
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
edge-tts==6.1.9
